new lateststatus

+++ status/lateststatus/mysql/entityone_ddl.go
@@ -14,16 +14,11 @@ func (link *Link) MigrateUp(ctx context.Context, exec sqlx.ExecerContext) (errEx
 	_, errExec = exec.ExecContext(
 		ctx,
 		`
-        CREATE TABLE IF NOT EXISTS entityone_status (
-            entityone_status_id BIGINT NOT NULL AUTO_INCREMENT,
-            action_id BIGINT NOT NULL DEFAULT 1,
-            status_id INT NOT NULL DEFAULT 1,
-            time_created DATETIME NOT NULL DEFAULT NOW(),
-            INDEX es_idx_sid (status_id ASC),
-            PRIMARY KEY (entityone_status_id)
-        )
-        ENGINE = InnoDB
-        DEFAULT CHARACTER SET = utf8;
+            CREATE TABLE IF NOT EXISTS entityone (
+                entityone_id BIGINT NOT NULL AUTO_INCREMENT,
+                time_created DATETIME NOT NULL DEFAULT NOW(),
+                PRIMARY KEY (entityone_id)
+            )
     `)
 	if errExec != nil {
 		return errExec
@@ -32,30 +27,56 @@ func (link *Link) MigrateUp(ctx context.Context, exec sqlx.ExecerContext) (errEx
 	_, errExec = exec.ExecContext(
 		ctx,
 		`
-        CREATE TABLE IF NOT EXISTS entityone (
-            entityone_id BIGINT NOT NULL AUTO_INCREMENT,
-            time_created DATETIME NOT NULL DEFAULT NOW(),
-            entityone_status_id BIGINT NOT NULL,
-            PRIMARY KEY (entityone_id),
-            UNIQUE INDEX e_idx_esi (entityone_status_id ASC),
-            CONSTRAINT e_fk_esi
-                FOREIGN KEY (entityone_status_id)
-                REFERENCES entityone_status (entityone_status_id)
-        )
-        ENGINE = InnoDB
-        DEFAULT CHARACTER SET = utf8
-        COLLATE = utf8_bin
+            CREATE TABLE IF NOT EXISTS entityone_status (
+                entityone_status_id BIGINT NOT NULL AUTO_INCREMENT,
+                entityone_id BIGINT NOT NULL,
+                action_id INT NOT NULL DEFAULT 1,
+                status_id INT NOT NULL DEFAULT 1,
+                time_created DATETIME NOT NULL DEFAULT NOW(),
+                PRIMARY KEY (entityone_status_id),
+                INDEX es_idx_sid (status_id),
+                INDEX es_fk_ei_e_idx (entityone_id),
+                CONSTRAINT es_fk_ei_e
+                    FOREIGN KEY (entityone_id)
+                    REFERENCES entityone (entityone_id)
+            )
+    `)
+	if errExec != nil {
+		return errExec
+	}
+
+	_, errExec = exec.ExecContext(
+		ctx,
+		`
+            CREATE TABLE IF NOT EXISTS entityone_lateststatus (
+                entityone_id BIGINT NOT NULL,
+                entityone_status_id BIGINT NOT NULL,
+                UNIQUE INDEX el_fk_es_esi_idx (entityone_status_id),
+                UNIQUE INDEX el_fk_e_ei_idx (entityone_id),
+                PRIMARY KEY (entityone_id, entityone_status_id),
+                CONSTRAINT el_fk_e_ei
+                    FOREIGN KEY (entityone_id)
+                    REFERENCES entityone (entityone_id),
+                CONSTRAINT el_fk_es_esi
+                    FOREIGN KEY (entityone_status_id)
+                    REFERENCES entityone_status (entityone_status_id)
+            )
     `)
 	return errExec
 }
 
 // MigrateDown destroys the needed tables
 func (link *Link) MigrateDown(ctx context.Context, exec sqlx.ExecerContext) (errExec error) {
-	_, errExec = exec.ExecContext(ctx, `DROP TABLE IF EXISTS entityone`)
+	_, errExec = exec.ExecContext(ctx, `DROP TABLE IF EXISTS entityone_lateststatus`)
 	if errExec != nil {
 		return errExec
 	}
 
 	_, errExec = exec.ExecContext(ctx, `DROP TABLE IF EXISTS entityone_status`)
+	if errExec != nil {
+		return errExec
+	}
+
+	_, errExec = exec.ExecContext(ctx, `DROP TABLE IF EXISTS entityone`)
 	return errExec
 }
